BUG/MEDIUM: stream: Prevent mux upgrades if client connection is no longer ready
If an early error occurred on the client connection, we must prevent any
multiplexer upgrades. Indeed, it is unexpected for a mux to be initialized
with no xprt. On a normal workflow it is impossible. So it is not an
issue. But if a mux upgrade is performed at the stream level, an early error
on the connection may have already been handled by the previous mux and the
connection may be already fully closed. If the mux upgrade is still
performed, a crash can be experienced.

It is possible to have a crash with an implicit TCP>HTTP upgrade if there is no
data in the input buffer. But it is also possible to get a crash with an
explicit "switch-mode http" rule.

It must be backported to all stable versions. In 2.2, the patch must be
applied directly in stream_set_backend() function.

(cherry picked from commit e4812404c541018ba521abf6573be92553ba7c53)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 13437097c312e524a346b9016d8ab273374d2053)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 56fb102c0c6094792fd38455b38b88a94454e996)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 2b48b04ab00148bc5ddfc66ba12214bdedf7c2ae)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit d36049a61ec1b5c081259018e50cb69d1452d99a)
[cf: si_conn_ready() is used instead of sc_conn_ready()]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>