BUG/MEDIUM: cli: Always release back endpoint between two commands on the mcli
When several commands are chained on the master CLI, the same client
connection is used. Because, it is a TCP connection, the mux PT is used. It
means there is no stream at the mux level. It is not possible to release the
applicative stream between each commands as for the HTTP. So, to work around
this limitation, between two commands, the master CLI is resetting the
stream. It does exactly what it was performed on HTTP to manage keep-alive
connections on old HAProxy versions.

But this part was copied from a code dealing with connection only while the
back endpoint can be an applet or a mux for the master cli. The previous fix
on the mux PT ("BUG/MEDIUM: mux-pt: Never fully close the connection on
shutdown") revealed a bug. Between two commands, the back endpoint was only
released if the connection's XPRT was closed. This works if the back
endpoint is an applet because there is no connection. But for commands sent
to a worker, a connection is used. At this stage, this only works if the
connection's XPRT is closed. Otherwise, the old endpoint is never detached
leading to undefined behavior on the next command execution (most probably a
crash).

Without the commit above, the connection's XPRT is always closed on
shutdown. It is no longer true. At this stage, we must inconditionnally
release the back endpoint by resetting the corresponding sedesc to fix the
bug.

This patch must be backported with the commit above in all stable
versions. On 2.4 and lower, it will need to be adapted.

(cherry picked from commit d4781bd5e7f0e0c0491fb7dfccb4da5c15055e3f)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 008f445a4fdbd1a56bafbf5fc863fe92e263a91f)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 14cd37372f099fe533e5f9296ce85f3bbf6940d6)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 1c2fb4cb2eddee33b8e770469fca1b69d3b70009)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 29e5b2695b0c4d620e46bf8576b8a0ede0ba7ab2)
[cf: The SI endpoint is always releaed]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>